Extract product rule tests into a dedicated file
Moved all test functions covering `product_rule_derivative`, `format_polynomial`, and `compute_polynomial_derivative_str` out from the monolithic `Tests/test_Calculus.py` into a newly created module-specific test file `Tests/test_product_rule.py`, complying with the project guidelines. Fixed previously shadowed test cases to ensure all written tests actually execute and provide complete coverage.
